+++ src/js/util.js
@@ -7,6 +7,7 @@ define(function () {
    * Parse JSON using the parser built-in in the browser.
    * On exception, the jsonString is validated and a detailed error is thrown.
    * @param {String} jsonString
+   * @return {JSON} json
    */
   util.parse = function parse(jsonString) {
     try {
@@ -15,7 +16,7 @@ define(function () {
     catch (err) {
       // try to load as JavaScript instead of JSON (like "{a: 2}" instead of "{"a": 2}"
       try {
-        return eval('(' + jsonString + ')');
+        return util.parseJS(jsonString);
       }
       catch(err2) {
         // ok no luck loading as JavaScript
@@ -27,6 +28,57 @@ define(function () {
         throw err;
       }
     }
+  };
+
+  /**
+   * Parse a string containing an object in JavaScript notation into a JSON.
+   * Throws an error when not successful. This function can for example parse
+   * a string like "{a: 2, 'b': {c: 'd'}".
+   * @param {string} jsString
+   * @returns {JSON} json
+   */
+  util.parseJS = function (jsString) {
+    // escape quotes inside strings
+    var chars = [];
+    var inString = false;
+    var i = 0;
+    while(i < jsString.length) {
+      var c = jsString.charAt(i);
+
+      if (c === '"' || c === '\'') {
+        if (c === inString) {
+          // end of string
+          inString = false;
+        }
+        else if (!inString) {
+          // start of string
+          inString = c;
+        }
+        else {
+          // add escape character if not escaped
+          if (jsString.charAt(i - 1) !== '\\') {
+            chars.push('\\');
+          }
+        }
+      }
+
+      chars.push(c);
+      i++;
+    }
+    var jsonString = chars.join('');
+
+    // replace keys/values enclosed by single quotes with double quotes
+    jsonString = jsonString.replace(/(.)'/g, function ($0, $1) {
+      var str = $1.replace(/"/g, '\\"'); // escape double quotes
+      return ($1 == '\\') ? '\'' : str + '"';
+    });
+
+    // enclose unquoted object keys with double quotes
+    jsonString = jsonString.replace(/([{,]\s*)([a-zA-Z_][a-zA-Z0-9_]*)(\s*:)/g, function ($0, $1, $2, $3) {
+      return $1 + '"' + $2 + '"' + $3;
+    });
+
+    return JSON.parse(jsonString);
   };
 
   /**